let's try with more attribute

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -3,4 +3,7 @@
 exec { "killmenow":
     command  => "pkill -f killmenow",
     path     => ['/bin', '/usr/bin', '/sbin', '/usr/sbin'],
+    provider => "shell",
+    onlyif   => "pgrep -f killmenow",
+    return_code => [0, 1],
 }
